Types and Constructors

Types and Constructors

In PopDyLan.jl, important data structures are given their own types. Thus, there is a type for a speaker, a type for a speech community, a type for a grammar, etc. These are referenced below, categorized into linguistic types (types for grammars, parameter settings, and the like), individual types (types for representing individual speakers) and population types (types for speech communities).

Linguistic Types

Main.PopDyLan.Variable — Type.

Variable

A linguistic variable; a probability distribution over a set of values.

source

Individual Types

Main.PopDyLan.Speaker — Type.

Speaker

source

Main.PopDyLan.VariationalLearner — Type.

VariationalLearner

A Variational Learner, a subtype of the abstract Speaker type.

source

Main.PopDyLan.VariationalLearner — Method.

VariationalLearner(g::Float64, A::Array{Float64,2}; age::Int=0, plastic::Bool=true)

Constructs a VariationalLearner with learning rate g and advantage matrix A.

The dimensions of A implicitly define the dimensionality of the learner. Note that no checking is performed to make sure that A is a valid advantage matrix. The learner's age and plasticity flag can be optionally set with the keyword arguments age and plastic. By default, each learner is constructed with age 0 and in a plastic state.

source

Main.PopDyLan.VariationalLearner — Method.

VariationalLearner(g::Float64, a1::Float64, a2::Float64; age::Int=0, plastic::Bool=true)

Constructs a classical 2-grammar (i.e. 1-parameter) VariationalLearner with learning rate g and advantage parameters a1 and a2.

The order of the advantage parameters is important: the first one becomes the advantage for grammar G1 (parameter on), the second one the advantage for grammar G2 (parameter off). The learner's age and plasticity flag can be optionally set with the keyword arguments age and plastic. By default, each learner is constructed with age 0 and in a plastic state.

source

Population Types

Main.PopDyLan.Community — Type.

Community

A Community is a collection of Speakers; technically, a Dict whose values are Speakers, with integer keys. A Community is empty when initialized; see inject! and eject!.

source