Methods

Methods

Methods (functions) operate on types, producing an output or modifying their inputs (or, sometimes, not doing anything at all). Names of methods that (may) modify their arguments end, per Julian conventions, in an exclamation point.

Note

In Julia, types do not own methods (like objects in, say, C++ do). As a consequence, to make foo act on bar, we do something like act(foo, bar) rather than foo.act(bar).

Methods are here divided into simulation methods (functions you use to conduct simulations), constructor methods (constructors for types) and auxiliary methods (e.g. functions for dealing with data).

Simulation Methods

Main.PopDyLan.getolder! — Method.

getolder!(x::Speaker)

Increase x's age by one.

source

Main.PopDyLan.listen! — Method.

listen!(x::VariationalLearner, y::Speaker, a)

Make Variational Learner x listen to Speaker y with grammatical advantages a. Learner x's state is (potentially) updated if the learner is plastic. The speaker's state is not modified.

source

Constructor Methods

Main.PopDyLan.Variable — Method.

Variable(n::Int)

Constructs a Variable of n values with a uniform distribution over those values.

source

Auxiliary Methods