Functions and Methods

Functions and Methods

Methods (functions) operate on types, producing an output or modifying their inputs (or, sometimes, not doing anything at all). Names of methods that (may) modify their arguments end, per Julian conventions, in an exclamation point.

Note

In Julia, types do not own methods (like objects in, say, C++ do). As a consequence, to make foo act on bar, we do something like act(foo, bar) rather than foo.act(bar).

Note

Constructor methods are documented under Types and Constructors.

Methods are here divided into simulation methods (functions you use to conduct simulations) and auxiliary methods (e.g. functions for dealing with data).

Simulation Methods

Main.PopDyLan.eject! — Function.

eject!(x::Int, y::Community)

Remove a Speaker with ID x from a Community.

source

Main.PopDyLan.getolder! — Function.

getolder!(x::Speaker)

Increase x's age by one.

source

getolder!(x::Community)

Increase the age of everyone in x by one.

source

Main.PopDyLan.inject! — Function.

inject!(x::Speaker, y::Community)

Add a Speaker to a Community.

source

Main.PopDyLan.learn! — Function.

learn!(x::VariationalLearner, y::Int)

Make VariationalLearner learn from input token y, which is simply a grammar index.

Called by listen!, this function is unlikely to be of use in itself to the end user. Implements Bush and Mosteller's Linear Reward-Penalty Learning algorithm, generalized for n dimensions.

source

Main.PopDyLan.listen! — Function.

listen!(x::VariationalLearner, y::VariationalLearner)

Make VariationalLearner x listen to one utterance from y and learn based on this input experience.

The state of x may change as a result; the state of y is guaranteed not to change.

source

Missing docstring.

Missing docstring for setprobabilities!. Check Documenter's build log for details.

Main.PopDyLan.speak — Function.

speak(x::VariationalLearner)

Causes VariationalLearner x to speak, i.e. to pick one of its grammars for production according to its internalized grammar probabilities. Returns an integer (the index of this grammar).

Called by listen!, this function is unlikely to be of use in itself to the end user.

source

Auxiliary Methods