Types

Types

In PopDyLan.jl, important data structures are given their own types. Thus, there is a type for a speaker, a type for a speech community, a type for a grammar, etc. These are referenced below, categorized into linguistic types (types for grammars, parameter settings, and the like), individual types (types for representing individual speakers) and population types (types for speech communities).

Linguistic Types

Main.PopDyLan.Variable — Type.

A linguistic variable; a probability distribution over a set of values.

source

Individual Types

Main.PopDyLan.VL — Type.

Variational Learner (1-parameter)

source

Population Types